Refactor code for more OOP
