+++ src/main/java/Ui.java
@@ -1,0 +1,27 @@
+import java.util.Scanner;
+
+public class Ui {
+    private Scanner scanner;
+    public Ui() {
+        this.scanner = new Scanner(System.in);
+    }
+    public String readCommand() {
+        return scanner.nextLine();
+    }
+
+    public void greet() {
+        System.out.println("Hello! I'm Bot");
+        System.out.println("What can I do for you?");
+    }
+
+    public void exit() {
+        System.out.println("Bye. Hope to see you again soon!");
+    }
+
+    public void listTasks(TaskList list) {
+        System.out.println("Here are the tasks in your list:");
+        for (int i = 0; i < list.size(); i++) {
+            System.out.println(i + 1 + ". " + list.get(i));
+        }
+    }
+}
